Update to mojo e2b1bc4e7cd4e3b6e851a45f96e55816bcc02457

+++ mojo/services/media/common/cpp/linear_transform.cc
@@ -306,8 +306,8 @@ bool LinearTransform::Ratio::Compose(const Ratio& a,
   MOJO_DCHECK(out);
   if (!out) { return false; }
 
-  int64_t  numerator = a.numerator * b.numerator;
-  uint64_t denominator = a.denominator * b.denominator;
+  int64_t  numerator   = static_cast<int64_t>(a.numerator) * b.numerator;
+  uint64_t denominator = static_cast<uint64_t>(a.denominator) * b.denominator;
 
   if (!numerator || !denominator) {
     out->numerator = numerator ? 1 : 0;
